Use radio button for selecting filter board
At this point, this is a purely cosmetic choice, but is meant to make it
simpler to add additional filter boards in the future. The current
(before this commit) design relied on interlocked check buttons that
emulated radio buttons with an additional off state, but doesn't
easily allow adding further filter board types.